Optimise signal filtering
Reduced the sample rate of the signal and tweak some features of the filtering algorithms to allow the use of SIMD instructions, also switching to single precision to make this more effective. This somehow broke SECAM decode a little, however. Also added a neatly formatted progress bar.

+++ VideoAnalogiser/PALSystem.h
@@ -39,8 +39,8 @@ private:
     FIRFilter colfir;
     FIRFilter lumaprefir;
     FIRFilter chromaprefir;
-    double* USignal;
-    double* VSignal;
-    double* USignalPreAlt;
-    double* VSignalPreAlt;
+    float* USignal;
+    float* VSignal;
+    float* USignalPreAlt;
+    float* VSignalPreAlt;
 };
